Contact_User_Code_Commit

+++ mycontact-backend/controllers/contactController.js
@@ -2,16 +2,16 @@ const asyncHandler = require('express-async-handler');
 const Contact = require("../models/contactModel")
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contact = await Contact.find();
+    const contact = await Contact.find({user_id: req.user.id});
     res.status(200).json(contact);
 });
 
 //@desc Create new contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
     console.log(req.body);
@@ -24,13 +24,14 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
 
     res.status(201).json(contact);
 });
 //@desc Get an individual contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
@@ -44,7 +45,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update  contacts
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
@@ -52,6 +53,12 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error ("Contact not Found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -62,7 +69,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete  contacts
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
@@ -70,6 +77,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
         throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
+
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
 });
